kmd_web/web_service/prodrequirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
Cartopy==0.25.0
celery==5.6.2
certifi==2026.2.25
cftime==1.6.5
click==8.3.1
gunicorn==25.1.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.12
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
fonttools==4.62.0
kiwisolver==1.5.0
kombu==5.6.2
matplotlib==3.10.8
netCDF4==1.7.4
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyparsing==3.3.2
pyproj==3.7.2
pyshp==3.0.3
python-dateutil==2.9.0.post0
redis==7.3.0
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.6.0
xarray==2026.2.0
